+++ src/components/DeleteDialog.js
@@ -1,22 +1,26 @@
 import React from 'react';
 
 const DeleteDialog = ({ isOpen, onClose, onDelete, title, message }) => {
-    if (!isOpen) return null;
+  if (!isOpen) return null;
 
-    return (
-        <div className="dialog-overlay">
-            <div className='dialog-content'>
-                <div className='dialog-title'>
-                    <h2>{title}</h2>
-                </div>
-                <p>{message}</p>
-                <div className='dialog-buttons'>
-                    <button className='cancel' onClick={onClose}>Cancel</button>
-                    <button className='delete' onClick={onDelete}>Delete</button>
-                </div>
-            </div>
+  return (
+    <div className="dialog-overlay">
+      <div className="dialog-content">
+        <div className="dialog-title">
+          <h4>{title}</h4>
         </div>
-    );
+        <p className="black">{message}</p>
+        <div className="dialog-buttons">
+          <button className="button nuetral" onClick={onClose}>
+            Cancel
+          </button>
+          <button className="button reject" onClick={onDelete}>
+            Delete
+          </button>
+        </div>
+      </div>
+    </div>
+  );
 };
 
 export default DeleteDialog;