+++ src/components/LogoutButton.js
@@ -1,21 +1,22 @@
-'use client'
-import { signOut } from 'next-auth/react'
-import { useRouter } from 'next/navigation'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import { faDoorOpen } from '@fortawesome/free-solid-svg-icons'
-import styles from './LogoutButton.module.css';
+'use client';
+import { signOut } from 'next-auth/react';
+import { useRouter } from 'next/navigation';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faDoorOpen } from '@fortawesome/free-solid-svg-icons';
 
 export default function LogoutButton() {
-    const router = useRouter()
+  const router = useRouter();
 
-    const handleLogout = async () => {
-        await signOut({ redirect: false })
-        router.push('/login')
-    }
+  const handleLogout = async () => {
+    await signOut({ redirect: false });
+    router.push('/login');
+  };
 
-    return (
-        <div className={styles.leftBox}>
-            <button onClick={handleLogout}><FontAwesomeIcon icon={faDoorOpen} /></button>
-        </div>
-    )
+  return (
+    <div className="pageHeaderLeft">
+      <button onClick={handleLogout} className="iconPrim">
+        <FontAwesomeIcon icon={faDoorOpen} />
+      </button>
+    </div>
+  );
 }